Made Navbar and moved Logout Logic to it
