+++ pokedex-react-flask-app/src/views/Home.js
@@ -3,15 +3,10 @@ import { useQuery } from "@apollo/client";
 import { GET_POKEMON_LIST } from "../api/pokeapi";
 import PokemonList from "../components/PokemonList";
 import { pokemonAPIClient } from "../api/clients";
-import { LOGOUT_MUTATION } from "../api/backend";
-import { backEndClient } from "../api/clients";
-import { useMutation } from "@apollo/client";
-import { useNavigate } from "react-router-dom";
+import NavBar from "../components/NavBar";
 
 export default function Home() {
-  const [user, setUser] = useState(JSON.parse(localStorage.getItem("user")));
-  const navigate = useNavigate();
-
+  const [user] = useState(JSON.parse(localStorage.getItem("user")));
   const [name, setName] = useState("");
   const { data: list, loading: loadingList } = useQuery(GET_POKEMON_LIST, {
     variables: { name: `%${name}%` },
@@ -25,33 +20,20 @@ export default function Home() {
     inputRef.current.focus();
   }, []);
 
-  const [logoutMutation] = useMutation(LOGOUT_MUTATION, {
-    client: backEndClient,
-  });
-
-  const handleLogout = async () => {
-    await logoutMutation();
-    localStorage.removeItem("token");
-    localStorage.removeItem("user");
-    setUser(null);
-    navigate("/login", { replace: true });
-  };
-
   return (
     <div className="App">
-      <header className="App-header">
-        <h2>Welcome {user?.name}</h2>
-        <button onClick={handleLogout}>Logout</button>
-        <p>Search For Pokemon</p>
-        <input
-          ref={inputRef}
-          value={name}
-          onChange={(val) => setName(val.target.value)}
-        ></input>
-        {!loadingList && list ? (
-          <PokemonList list={list.pokemon_list} />
-        ) : undefined}
-      </header>
+      <NavBar />
+      <h2>Welcome {user?.name}</h2>
+
+      <p>Search For Pokemon</p>
+      <input
+        ref={inputRef}
+        value={name}
+        onChange={(val) => setName(val.target.value)}
+      ></input>
+      {!loadingList && list ? (
+        <PokemonList list={list.pokemon_list} />
+      ) : undefined}
     </div>
   );
 }
